// src/shared/api/generated/model/complex.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * Display Management API
 * API для управления домами, дисплеями, группами, шаблонами и экстренными сообщениями.
 *
 * ## Аутентификация
 * Все эндпоинты (кроме `/auth/*`) требуют JWT Bearer токен в заголовке:
 * ```
 * Authorization: Bearer <accessToken>
 * ```
 *
 * OpenAPI spec version: 2.0.0
 */
import type { House } from './house';

export interface Complex {
  id: number;
  name: string;
  houses: House[];
}